disploy 7/8

+++ lib/features/seller/dashboard/data/models/revenue_model.dart
@@ -6,7 +6,7 @@ class RevenueModel extends RevenueEntity {
   @override
   RevenueModel copyWith({
     String? key,
-    double? value,
+    int? value,
   }) =>
       RevenueModel(
         key: key ?? this.key,
